Pixelated Lighting
P5 render now has a convincing lighting look just by interpolating between the red and green channels of the loaded normal map. The scale is not consistent with the canvas and the loop is nowhere near seamless. Nevertheless, this iteration is a good proof of concept and gives me confidence that this design should be very effective.